pyproject.toml
[tool.poetry]
name = "rag-system"
version = "0.1.0"
description = "Production-grade RAG system with conversational capabilities and graph-augmented retrieval for ArXiv papers"
authors = ["RAG System Team"]
readme = "README.md"
packages = [{include = "rag_system"}]

[tool.poetry.dependencies]
python = "^3.10"
arxiv = "^2.1.0"
pymupdf = "^1.24.0"
networkx = "^3.2"
chromadb = "^0.4.22"
sentence-transformers = "^2.2.2"
rank-bm25 = "^0.2.2"
fastapi = "^0.109.0"
uvicorn = {extras = ["standard"], version = "^0.27.0"}
gradio = "^4.15.0"
httpx = "^0.26.0"
pydantic = "^2.5.0"
python-dotenv = "^1.0.0"
numpy = "^1.26.0"
scipy = "^1.12.0"
scikit-learn = "^1.4.0"
torch = "^2.1.0"
transformers = "^4.37.0"
ragas = "^0.1.0"
datasets = "^2.16.0"
plotly = "^5.18.0"
python-community = {git = "https://github.com/taynaud/python-louvain.git", optional = true}
groq = "^0.4.0"
openai = "^1.10.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
ruff = "^0.2.0"
mypy = "^1.8.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "W"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true
